chore: moved gallery data to config
