+++ src/config/gallery/index.js
